rename spec -> interface
