+++ lib/hackmode/index.js
@@ -1,6 +1,6 @@
 const msgpack = require('msgpack-lite')
 
-const { P2p } = require('../n3h-mod-spec')
+const { P2p } = require('../interface')
 const { P2pBackendHackmodePeer } = require('./p2p-backend-hackmode-peer')
 
 const { N3hMode } = require('../n3h-ipc/n3hMode')
