+++ lib/hackmode/p2p-backend-hackmode-peer.js
@@ -14,7 +14,7 @@ const {
   Dht,
   DhtEvent,
   P2pEvent
-} = require('../n3h-mod-spec')
+} = require('../interface')
 
 const { ConnectionBackendWss } = require('../n3h-mod-connection-wss')
 const { DhtBackendFullsync } = require('../n3h-mod-dht-fullsync')
