+++ lib/hackmode/p2p-backend-hackmode-peer.test.js
@@ -11,7 +11,7 @@ const tmp = require('tmp-promise')
 
 const { expect } = require('chai')
 
-const { P2p } = require('../n3h-mod-spec')
+const { P2p } = require('../interface')
 const { P2pBackendHackmodePeer } = require('./p2p-backend-hackmode-peer')
 
 describe('hackmode module p2p peer backend Suite', () => {
