+++ lib/hackmode/realmode.js
@@ -1,6 +1,5 @@
 const msgpack = require('msgpack-lite')
 
-const { P2p } = require('../n3h-mod-spec/index')
 const { P2pBackendHackmodePeer } = require('../hackmode/p2p-backend-hackmode-peer')
 
 const { N3hMode } = require('../n3h-ipc/n3hMode')
@@ -13,9 +12,10 @@ const log = tweetlog('@realmode@')
 const { DhtBackendFullsync } = require('../n3h-mod-dht-fullsync/index')
 
 const {
+  P2p,
   Dht,
   DhtEvent
-} = require('../n3h-mod-spec/index')
+} = require('../interface')
 
 const { $sleep } = require('../n3h-common')
