+++ lib/interface/connection-iface.js
