+++ lib/interface/index.js
@@ -10,6 +10,6 @@ function exportReq (mod) {
   }
 }
 
-exportReq(require('./connection-spec'))
-exportReq(require('./dht-spec'))
-exportReq(require('./p2p-spec'))
+exportReq(require('./connection-iface'))
+exportReq(require('./dht-iface'))
+exportReq(require('./p2p-iface'))
