+++ lib/n3h-ipc/n3hMode.js
@@ -2,7 +2,7 @@ const msgpack = require('msgpack-lite')
 
 const { AsyncClass } = require('../n3h-common')
 
-const { Connection } = require('../n3h-mod-spec')
+const { Connection } = require('../interface')
 const { ConnectionBackendWss } = require('../n3h-mod-connection-wss')
 
 const config = require('../hackmode/config')
