+++ lib/n3h-mod-connection-wss/wss-connection.js
@@ -5,7 +5,7 @@ const Wss = require('ws')
 
 const { AsyncClass } = require('../n3h-common')
 const { WssServer } = require('./wss-server')
-const { ConnectionEvent } = require('../n3h-mod-spec')
+const { ConnectionEvent } = require('../interface')
 
 const tweetlog = require('../tweetlog')
 const log = tweetlog('wss-connection')
