+++ lib/n3h-mod-connection-wss/wss-connection.test.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai')
-const { Connection } = require('../n3h-mod-spec')
+const { Connection } = require('../interface')
 const { ConnectionBackendWss } = require('./index')
 const { $sleep } = require('../n3h-common')
