+++ lib/n3h-mod-dht-fullsync/dht-fullsync.js
@@ -7,7 +7,7 @@ const {
   Executor,
   $sleep
 } = require('../n3h-common')
-const { Dht } = require('../n3h-mod-spec')
+const { Dht } = require('../interface')
 const DhtEvent = Dht.DhtEvent
 
 const tweetlog = require('../tweetlog')
