+++ lib/n3h-mod-dht-fullsync/dht-fullsync.test.js
@@ -1,6 +1,6 @@
 const { expect } = require('chai')
 
-const { Dht, DhtEvent } = require('../n3h-mod-spec')
+const { Dht, DhtEvent } = require('../interface')
 const { DhtBackendFullsync } = require('./index')
 
 const TEST_URI_1 = 'fake://some/place'
